spring-jpa/src/main/java/com/ashan/example/spring/jpa/service/EmployeeService.java
package com.ashan.example.spring.jpa.service;

import com.ashan.example.spring.jpa.entity.Employee;
import com.ashan.example.spring.jpa.repository.EmployeeRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@AllArgsConstructor
public class EmployeeService {

    private EmployeeRepository repository;

    public Employee saveEmployee(Employee employee) {
        return repository.save(employee);
    }

    public List<Employee> getAllEmployee() {
        return repository.get();
    }
}
